The main file in the Gateway project
Becase all the project is too big. If you want to run the project, you have to install all of them. So, I upload a main file here if you want to preview before install. The principle here is: First connect and subcribe to feeds on Adafruit then receive message from output and write them feeds corresponding (Value is a string has the json form)